cast double -> float converion
git-svn-id: file:///fltk/svn/trunk@3495 ea41ed52-d2ee-0310-a9c1-e6b18d33e121

+++ glut/trackball.c
@@ -113,7 +113,7 @@ vcross(const float *v1, const float *v2, float *cross)
 float
 vlength(const float *v)
 {
-    return sqrt(v[0] * v[0] + v[1] * v[1] + v[2] * v[2]);
+    return (float)(sqrt(v[0] * v[0] + v[1] * v[1] + v[2] * v[2]));
 }
 
 void
